Change `size_t` to `int`

+++ examples/depnodep.c
@@ -2,7 +2,7 @@
 #include <stdlib.h>
 
 void
-print_arr(size_t N, double *A) {
+print_arr(int N, double *A) {
   for (int i = 0; i < N; ++i) {
     printf("%lf, ", A[i]);
   }
@@ -10,19 +10,19 @@ print_arr(size_t N, double *A) {
 }
 
 void
-init_arr(size_t N, double *A) {
+init_arr(int N, double *A) {
   for (int i = 0; i < N; ++i) {
     A[i] = random();
   }
 }
 
-const size_t N = 10;
+const int N = 10;
 
 void
-f(size_t N, double A[N][N]) {
+f(int N, double A[N][N]) {
 #pragma scop
-  for (size_t i = 1; i < N; i++) {
-    for (size_t j = 0; j < N; j++) {
+  for (int i = 1; i < N; i++) {
+    for (int j = 0; j < N; j++) {
       A[i][j] = A[i][j] + A[i - 1][j];
     }
   }
